chore: refactor custom components
