--- libs/application/templates/public-debt-payment-plan/src/fields/EmployerInfo/EmployerInfoDescription.css.ts
@@ -1,5 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-export const wrapper = style({
-  marginTop: -24,
-})
