+++ libs/application/templates/public-debt-payment-plan/src/fields/index.ts
@@ -1,7 +1,4 @@
 export * from './DisposableIncome/DisposableIncome'
-export * from './EmployerInfo/EmployerInfo'
-export * from './EmployerInfo/EmployerInfoDescription'
 export * from './Overview/Overview'
 export * from './PaymentPlan/PaymentPlan'
 export * from './PaymentPlanList/PaymentPlanList'
-export * from './DescriptionWithLink'
